+++ zp_sitl/sil_drivers/sil_queue.hpp
@@ -25,10 +25,10 @@ public:
     }
     
     int count() override {
-        return q.size();
+        return (int)q.size();
     }
     
     int remainingCapacity() override {
-        return maxSize - q.size();
+        return (int)(maxSize - q.size());
     }
 };
